src/main/java/com/cts/projectmanager/dto/UserDTO.java
package com.cts.projectmanager.dto;

import java.io.Serializable;

/**
 * User DTO class for project Management.
 * @author Ramakrishna Gurram
 */
public class UserDTO implements Serializable {
    private int userId;
    private String firstName;
    private String lastName;
    private int employeeId;

    public int getUserId() {
        return userId;
    }

    public void setUserId(int userId) {
        this.userId = userId;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public int getEmployeeId() {
        return employeeId;
    }

    public void setEmployeeId(int employeeId) {
        this.employeeId = employeeId;
    }

    @Override
    public String toString() {
        return "UserDTO{" +
                "userId=" + userId +
                ", firstName='" + firstName + '\'' +
                ", lastName='" + lastName + '\'' +
                ", employeeId=" + employeeId +
                '}';
    }
}
